derivadas primeira com erros de ordens 2 e 3 por série de taylor

+++ include/function.h
@@ -3,6 +3,6 @@
 
 #include <cmath>
 
-float F(float x);
+double F(double x);
 
 #endif
